# specs/001-todo-ai-chatbot/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Todo AI Chatbot API
  description: API for the AI-powered todo management chatbot
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
paths:
  /api/{user_id}/chat:
    post:
      summary: Send message and get AI response
      parameters:
        - name: user_id
          in: path
          required: true
          description: The user's identifier
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                conversation_id:
                  type: integer
                  description: Existing conversation ID (creates new if not provided)
                  example: 123
                message:
                  type: string
                  description: User's natural language message
                  example: "Add a task to buy groceries"
              required:
                - message
      responses:
        '200':
          description: AI assistant response
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation_id:
                    type: integer
                    description: The conversation ID
                    example: 123
                  response:
                    type: string
                    description: AI assistant's response
                    example: "I've added the task 'buy groceries' for you."
                  tool_calls:
                    type: array
                    description: List of MCP tools invoked
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                          description: Name of the tool called
                          example: "add_task"
                        parameters:
                          type: object
                          description: Parameters passed to the tool
                          example: {"user_id": "user123", "title": "buy groceries"}
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /health:
    get:
      summary: Health check endpoint
      responses:
        '200':
          description: Health status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"